fixes: runtime parity fixes (no tests)

+++ src/eghudm.c
@@ -42,7 +42,7 @@ int FAR CDECL hudSine(int angle) { return nsine(angle); }
 
 /* pitch -> ladder pixel offset: (|pitch|>>6) * 360, taking the high bytes. */
 int FAR CDECL hudPitchScale(int ap) {
-    return (int)(((unsigned long)(unsigned)ap * 360u) >> 8);
+    return (int)(((unsigned long)(uint16)ap * 360u) >> 8);
 }
 
 /* ===== drawClipLineGlobal (clipLineFar) =====
